Align GitHub source control adapter errors
